Major Reorganization
After working with this library for awhile I'm starting
to realize that lib.rs is getting too big. So I've
reoganized the code into modules to make it more
natural to use the library. Everything relating to
Volumes goes in volume.rs, everything relating to Peers
goes in peers.rs, etc.